# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Orange Money API
  description: |
    API REST complète pour la gestion des transactions mobiles Orange Money.
    
    ## Fonctionnalités principales
    - 💰 **Transactions** : Dépôts, retraits, transferts, paiements
    - 👥 **Gestion utilisateurs** : Clients, distributeurs, marchands, admin
    - 💳 **Gestion comptes** : Consultation solde, historique, blocage/déblocage
    - 🔐 **Authentification JWT** : Tokens sécurisés avec cookies HTTP-only
    
    ## Authentification
    1. Utilisez l'endpoint `/api/auth/login` pour vous connecter
    2. Copiez le token depuis la réponse (il est aussi dans un cookie HTTP-only)
    3. Cliquez sur "Authorize" et collez le token
    4. Testez les endpoints protégés
    
    ## Données de test (utilisateurs sans code PIN)
    | Rôle | Nom | Téléphone | Mot de passe | Numéro de compte |
    |------|-----|-----------|--------------|------------------|
    | CLIENT | Moussa Diop | 771234567 | Password123! | OM8000380279 |
    | CLIENT | Fatou Sarr | 779876543 | Pass123! | OM2665616523 |
    | CLIENT | Cheikh Fall | 776543210 | Client123! | OM5432147504 |
    | DISTRIBUTEUR | Abdou Ndiaye | 775551234 | Distrib123! | OM4274060223 |
    

  version: 1.0.0
  contact:
    name: Orange Money Support
    email: support@orangemoney.sn
    url: https://www.orangemoney.sn
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8083
    description: Serveur de développement
  - url: https://api.orangemoney.sn
    description: Serveur de production

tags:
  - name: Authentification
    description: Gestion de l'authentification et des sessions utilisateur
  - name: Utilisateurs
    description: Gestion des utilisateurs (CRUD)
  - name: Comptes
    description: Gestion des comptes bancaires
  - name: Transactions
    description: Opérations de dépôt, retrait, transfert et paiement

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu après login.
        Le token est automatiquement stocké dans un cookie HTTP-only.
        Pour Swagger, copiez le token depuis la réponse du login.

  schemas:
    RegisterRequest:
      type: object
      required:
        - nom
        - prenom
        - telephone
        - email
        - motDePasse
        - codePin
        - role
      properties:
        nom:
          type: string
          example: Diop
        prenom:
          type: string
          example: Moussa
        telephone:
          type: string
          pattern: '^(77|78|76|70|75)[0-9]{7}$'
          example: '771234567'
        email:
          type: string
          format: email
          example: moussa@test.com
        motDePasse:
          type: string
          format: password
          example: Password123!
        codePin:
          type: string
          pattern: '^[0-9]{6}$'
          example: '482915'
        role:
          type: string
          enum: [CLIENT, DISTRIBUTEUR, MARCHAND, ADMIN]
          example: CLIENT

    LoginRequest:
      type: object
      required:
        - telephone
        - motDePasse
      properties:
        telephone:
          type: string
          example: '771234567'
        motDePasse:
          type: string
          format: password
          example: Password123!

    TransfertRequest:
      type: object
      required:
        - telephoneDestinataire
        - montant
      properties:
        telephoneDestinataire:
          type: string
          example: '779876543'
        montant:
          type: number
          format: double
          minimum: 100
          maximum: 1000000
          example: 5000

    DepotRequest:
      type: object
      required:
        - telephoneClient
        - montant
      properties:
        telephoneClient:
          type: string
          example: '771234567'
        montant:
          type: number
          format: double
          example: 50000

    RetraitRequest:
      type: object
      required:
        - telephoneClient
        - montant
      properties:
        telephoneClient:
          type: string
          example: '771234567'
        montant:
          type: number
          format: double
          example: 10000

    PaiementRequest:
      type: object
      required:
        - codeMarchand
        - montant
      properties:
        codeMarchand:
          type: string
          example: SHOP001
        montant:
          type: number
          format: double
          example: 15000

    AuthResponse:
      type: object
      properties:
        type:
          type: string
          example: Bearer
        telephone:
          type: string
          example: '771234567'
        nom:
          type: string
          example: Diop
        prenom:
          type: string
          example: Moussa
        role:
          type: string
          example: CLIENT

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        reference:
          type: string
          example: TRXCC41124F
        typeTransaction:
          type: string
          enum: [DEPOT, RETRAIT, TRANSFERT, PAIEMENT]
          example: TRANSFERT
        montant:
          type: number
          format: double
          example: 5000.0
        frais:
          type: number
          format: double
          example: 100.0
        montantTotal:
          type: number
          format: double
          example: 5100.0
        statut:
          type: string
          enum: [EN_ATTENTE, REUSSI, ECHOUE, ANNULE]
          example: REUSSI
        nouveauSolde:
          type: number
          format: double
          description: Nouveau solde du compte expéditeur après la transaction
          example: 44900.0

    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Opération réussie
        data:
          type: object
        timestamp:
          type: string
          format: date-time

paths:
  /api/auth/register:
    post:
      tags:
        - Authentification
      summary: Inscription d'un nouvel utilisateur
      description: Créer un nouveau compte utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Inscription réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/login:
    post:
      tags:
        - Authentification
      summary: Connexion utilisateur
      description: Authentifier un utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/auth/logout:
    post:
      tags:
        - Authentification
      summary: Déconnexion
      description: Déconnecter l'utilisateur
      responses:
        '200':
          description: Déconnexion réussie

  /api/auth/me:
    get:
      tags:
        - Authentification
      summary: Utilisateur connecté
      description: Informations de l'utilisateur connecté
      responses:
        '200':
          description: Informations récupérées

  /api/transactions/transfert:
    post:
      tags:
        - Transactions
      summary: Transfert d'argent
      description: |
        Transférer de l'argent entre deux clients.
        **Frais**: 0.85% du montant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransfertRequest'
      responses:
        '200':
          description: Transfert réussi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/transactions/depot:
    post:
      tags:
        - Transactions
      summary: Dépôt d'argent
      description: |
        Déposer de l'argent (DISTRIBUTEUR uniquement).
        **Frais**: 0 FCFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepotRequest'
      responses:
        '200':
          description: Dépôt réussi

  /api/transactions/retrait:
    post:
      tags:
        - Transactions
      summary: Retrait d'argent
      description: |
        Retirer de l'argent via distributeur.
        **Frais**: 0 FCFA (aucun frais)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetraitRequest'
      responses:
        '200':
          description: Retrait réussi

  /api/transactions/paiement:
    post:
      tags:
        - Transactions
      summary: Paiement marchand
      description: |
        Payer un marchand.
        **Frais**: 0.85% du montant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaiementRequest'
      responses:
        '200':
          description: Paiement réussi

  /api/transactions/historique/{numeroCompte}:
    get:
      tags:
        - Transactions
      summary: Historique des transactions
      description: Historique complet
      parameters:
        - name: numeroCompte
          in: path
          required: true
          schema:
            type: string
            example: OM8000380279
      responses:
        '200':
          description: Historique récupéré

  /api/transactions/historique/{numeroCompte}/periode:
    get:
      tags:
        - Transactions
      summary: Historique par période
      parameters:
        - name: numeroCompte
          in: path
          required: true
          schema:
            type: string
        - name: dateDebut
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-11-08T00:00:00'
        - name: dateFin
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-11-08T23:59:59'
      responses:
        '200':
          description: Historique récupéré

  /api/transactions/annuler/{reference}:
    post:
      tags:
        - Transactions
      summary: Annuler une transaction
      description: |
        Annuler un transfert ou paiement effectué.
        Seul l'expéditeur peut annuler sa propre transaction.
        Le montant + frais sont remboursés automatiquement.
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
            example: TRXABC12345
          description: Référence de la transaction à annuler
      responses:
        '200':
          description: Transaction annulée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Transaction déjà annulée ou non annulable
        '401':
          description: Vous ne pouvez annuler que vos propres transactions
        '404':
          description: Transaction non trouvée

  /api/comptes/solde/{numeroCompte}:
    get:
      tags:
        - Comptes
      summary: Consulter le solde
      parameters:
        - name: numeroCompte
          in: path
          required: true
          schema:
            type: string
            example: OM8000380279
      responses:
        '200':
          description: Solde récupéré

  /api/comptes/utilisateur/{utilisateurId}:
    get:
      tags:
        - Comptes
      summary: Comptes d'un utilisateur
      parameters:
        - name: utilisateurId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste des comptes